requirements.txt
newspaper3k
googletrans
tkinter